Rename TestUtils.ts to TestUtils.test.ts
This will allow it to be excluded when we do #863
