+++ src/InputHandler.test.ts
@@ -5,7 +5,7 @@
 import { assert } from 'chai';
 import { InputHandler } from './InputHandler';
 import { wcwidth } from './InputHandler';
-import { MockInputHandlingTerminal } from './utils/TestUtils';
+import { MockInputHandlingTerminal } from './utils/TestUtils.test';
 
 describe('InputHandler', () => {
   describe('save and restore cursor', () => {
