+++ src/SelectionModel.test.ts
@@ -6,7 +6,7 @@ import { assert } from 'chai';
 import { ITerminal } from './Interfaces';
 import { SelectionModel } from './SelectionModel';
 import {BufferSet} from './BufferSet';
-import { MockTerminal } from './utils/TestUtils';
+import { MockTerminal } from './utils/TestUtils.test';
 
 class TestSelectionModel extends SelectionModel {
   constructor(
