+++ src/Terminal.test.ts
@@ -1,6 +1,6 @@
 import { assert, expect } from 'chai';
 import { Terminal } from './Terminal';
-import { MockViewport, MockCompositionHelper } from './utils/TestUtils';
+import { MockViewport, MockCompositionHelper } from './utils/TestUtils.test';
 import { CHAR_DATA_CHAR_INDEX, CHAR_DATA_WIDTH_INDEX } from './Buffer';
 
 class TestTerminal extends Terminal {
